Add usage, verbose option, helptext. Add arm64 arch.

+++ download-terraform.sh
@@ -8,6 +8,28 @@
 
 cd "$(mktemp -d)" || exit 1
 
+usage(){ cat <<EOF
+Download several versions of terraform into ~/bin
+
+    -h  Print help
+    -v  Enable verbose
+    -x  Enable xtrace
+
+${0##*/} [-x] [version_number ... version_number]
+EOF
+}
+
+extra_args=()
+
+while getopts ':xv' option ; do
+  case "${option}" in
+    x) set -x ;;
+    v) export VERBOSE=1 ; extra_args+=( -w '%{url_effective}\n' ) ;;
+    *) usage ; exit 1 ;
+  esac
+done
+shift $((OPTIND - 1))
+
 if [[ -z "${*}" ]] ; then
   versions=(
     "0.12.31" # 2021-04-26
@@ -35,10 +57,20 @@ download_version() {
     armv5*) arch="armv5" ;;
     armv6*) arch="armv6" ;;
     armv7*) arch="armv7" ;;
+    arm64*) arch="arm64" ;;
     *) echo "ABORT: Unknown architecture '$arch'" ; exit 1 ;;
   esac
 
-  curl -sSLfo terraform.zip "https://releases.hashicorp.com/terraform/${version}/terraform_${version}_${os,,}_${arch}.zip" || return
+  [ -n "$VERBOSE" ] && { echo ; echo "Trying to download with os=$os, arch=$arch, version=$version" ; }
+
+  URL="https://releases.hashicorp.com/terraform/${version}/terraform_${version}_${os,,}_${arch}.zip"
+
+  if ! curl -sSLf "${extra_args[@]}" -o terraform.zip "$URL" ; then
+    echo "Failed to download $URL"
+    return
+  fi
+
+
   unzip terraform.zip >/dev/null
   rm -rf terraform.zip
   mv terraform "${dest}"
